Backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
